Add about page, make footer compatible with small pages
The new About page adds my resume, and more details about myself, to the website.

The new footer pins the footer to the bottom of the page if a small page is designated, and sticks on the end of content on large pages.

An error exists in the new footer on thin screens where the footer covers page content. The fix is being delayed in favor of a redesign using the Hugo framework