:tada: Release figbert.com version 2
The version 2 rework of figbert.com comes after months of development, and is a total rework of the site. figbert.com has moved from Hugo with the Coder theme to a custom design implemented with Sapper and TailwindCSS. The new version is a massive upgrade, and will be further detailed in an upcoming blog post.