:art: Move emojis from img tags to inline svg
To reduce the number of requests to the server, the redesigned emoji shortcodes embed the emoji svg directly into the HTML instead of through an img tag.

Additionally, a `height: auto;` element has been added to the emoji class so that it is clearly defined and doesn't create unforseen CSS problems.