Redesign /reading
The Current and To-Read subsections have been removed, under the
philosophy that committing to uncertain actions before they happen is
less good than simply publishing the actions as they are taken.

The very-similar templates for posts/projects and reading have further
been merged into one template for ease of maintenance.